aws-msp-funding-stack/aws-cur-parser-lambda/requirements.txt
boto3
pyyaml
datetime